move utils to metadata
